shim voiceMove to trim 30k off round.min.js
