+++ ui/round/src/view/main.ts
@@ -48,7 +48,7 @@ export function main(ctrl: RoundController): VNode {
           },
           [renderGround(ctrl), ctrl.promotion.view(ctrl.data.game.variant.key === 'antichess')]
         ),
-        ctrl.voiceMove ? renderVoiceMove(ctrl.voiceMove, false) : null,
+        ctrl.voiceMove ? renderVoiceMove(ctrl.redraw, false) : null,
         ctrl.keyboardHelp ? keyboard.view(ctrl) : null,
         crazyView(ctrl, topColor, 'top') || materialDiffs[0],
         ...renderTable(ctrl),
